Filters: create the driver definitions
Currently the architecture for an S(3) filter and a generic filter
driver.  These drivers contain the callbacks to the particular
implementation.  This will still likely be interface code as there
will be steps check if the data is sufficient to run the filter
etc that is separate for each filter math.  Therefore it is proposed
most filters be implemented as three files

filter_name_core.c       -- contains the core math for the filter
filter_name_interface.c  -- contains the code mapping from the filter class
                         -- API defined here to the core.  Also defines the
                         -- driver as a flash constant instance
filter_name_interface.h  -- contains the extern define of the filter driver