QCINS: mag updates only attempt to update quaternion
In addition to only projecting the change from the mag
along the heading axis, this also only updates the quaternion
state. This prevents these updates influencing the other states
and also applies to the covariance updates.

In the future, it might make sense to reproject some of the
corrections for rates and torques (maybe even biases) into
the body frame and apply a correction to that.